# src/main/resources/api-specs/customer-api-specs.yml
openapi: 3.0.3
info:
  title: Product API
  description: API for managing products in an online shop
  version: 1.0.0

paths:
  /api/products:
    get:
      summary: Get all products
      tags:
        - Products
      responses:
        '200':
          description: List of available products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorResponse" }
              example:
                status: 500
                error: "Internal Server Error"
                message: "Unexpected error occurred"
                path: "/api/products"
    post:
      summary: Create a new product
      tags:
        - Products
      requestBody:
        description: Product data to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '200':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict — product with same unique attribute already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorResponse" }
              example:
                status: 409
                error: "Conflict"
                message: "Product with this name already exists"
                path: "/api/products"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorResponse" }
              example:
                status: 500
                error: "Internal Server Error"
                message: "Unexpected error occurred"
                path: "/api/products"

  /api/products/{id}:
    get:
      summary: Get a product by id
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the product
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorResponse" }

    put:
      summary: Update a product by id
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the product
      requestBody:
        description: Updated product data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorResponse" }

    delete:
      summary: Delete a product by id
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the product
      responses:
        '204':
          description: Product deleted successfully (no content)
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorResponse" }

components:
  schemas:
    ProductDto:
      type: object
      required:
        - name
        - description
        - price
        - category
      properties:
        name:
          type: string
          maxLength: 100
          description: Name of the product
        description:
          type: string
          description: Description of the product
        price:
          type: number
          description: Price of the product
        category:
          type: object
          required:
            - id
            - name
          properties:
            id:
              type: integer
              format: int64
              description: Category ID
            name:
              type: string
              description: Category name

    ApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: HTTP status reason phrase
          example: Not Found
        message:
          type: string
          description: Detailed error message
          example: Product with ID 123 not found
        path:
          type: string
          description: Request path that caused the error
          example: /api/products/123